-feat : 레스토랑 메뉴 찾지 못할 때 예외처리 코드 추가
